requirements.txt
streamlit>=1.42.0
pycryptodome>=3.21.0
